Updated Photon and VRSF Scripts based on the test made in another multiplayer project

+++ VRSF_DOTS/Assets/VRSF/Core/Scripts/Utils/Systems/DestroyVRSFEntities.cs
@@ -5,10 +5,14 @@ namespace VRSF.Core.Utils
 {
     public class EntitiesDestroyOnSceneLoaded : ComponentSystem
     {
+        private string _activeScene;
+
         protected override void OnCreate()
         {
             base.OnCreate();
             SceneManager.sceneUnloaded += DestroyEntities;
+            SceneManager.sceneLoaded += SetActiveScene;
+            _activeScene = SceneManager.GetActiveScene().name;
         }
 
         protected override void OnUpdate() { }
@@ -17,14 +21,29 @@ namespace VRSF.Core.Utils
         {
             base.OnDestroy();
             SceneManager.sceneUnloaded -= DestroyEntities;
+            SceneManager.sceneLoaded -= SetActiveScene;
         }
 
         private void DestroyEntities(Scene unloadedScene)
         {
+            if (unloadedScene.name != _activeScene)
+                return;
+
             Entities.ForEach((Entity e, ref DestroyOnSceneUnloaded destroyComponent) =>
             {
                 EntityManager.DestroyEntity(e);
             });
+
+            Entities.ForEach((Entity e, ref DestroyOnSceneUnloaded destroyComponent, ref Disabled disabledEntityComp) =>
+            {
+                EntityManager.DestroyEntity(e);
+            });
+        }
+
+        private void SetActiveScene(Scene loadedScene, LoadSceneMode loadMode)
+        {
+            if (loadMode == LoadSceneMode.Single)
+                _activeScene = loadedScene.name;
         }
     }
 }
